test: isolate non-Unix network quit coverage
